FEAT: rewrite to support general VMM
By default functions vir_mem_alloc and, vir_mem_free will not map any
memory it is just going to reserv the virtual memory the caller is
responsable of mapping.

For ease of use the function get_physical is provided.

When mapping pages the physical and virtual size must match if not the
kernel will panic

kmalloc functions are rm from the vir_mem.c